requirements.txt
ffmpeg-python
speechbrain
soundfile
torch
pydub
openai-whisper
pyannote.core
pyannote.audio
tqdm
demucs